pyproject.toml
[project]
name = "aws-test"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "nao2c4", email = "27849770+nao2c4@users.noreply.github.com" }
]
requires-python = ">=3.13"
dependencies = [
    "pydantic>=2.10.6",
    "pyyaml>=6.0.2",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[dependency-groups]
dev = [
    "ipykernel>=6.29.5",
    "pre-commit>=4.1.0",
    "pytest>=8.3.5",
    "ruff>=0.9.10",
]

[tool.ruff]
# Group violations by containing file.
output-format = "grouped"
# Allow lines to be as long as 100.
line-length = 100

[tool.ruff.lint]
select = ["ALL"]
ignore = [
    "COM812",  # [[conflicted]]
    "ISC001",  # [[conflicted]]
    "D415",    # Period
]

[tool.ruff.lint.pydocstyle]
convention = "google"

[tool.ruff.lint.per-file-ignores]
"tests/**.py" = ["S101"]

[[tool.mypy.overrides]]
module = ["scipy.*"]
ignore_missing_imports = true

[tool.pytest.ini_options]
minversion = "6.0"
addopts = "-ra -q"
testpaths = ["tests"]
